Update all strings in po files.
The translators need the .po files updated with all of the current
strings.  It is not sufficient to update just the anaconda.pot file for
the translation team.  The only time the pot file is used is when a new
language is added.  If a language exists, the translators only look at
the po file and expect to see all strings there.