Remove PartitionDevice.path hack.
This was put in place for fear that non-existent partitions might cause
multiple PartitionDevice instances to have the same value for their path
property. The one thing is certainly does is make it so that non-existent
partitions' path is not actually a path, which means we've chosen the wrong
way around some other problem.