mpath: add MultipathDevices first, before their partitions.
Otherwise the partitions will be ignored because their disks (the mpath
devices) don't have disklabels yet.

Also in udev_device_get_uuid() compare just the plain strings, not the
punctuation. This resulted in some of the devices with remmants of MD raid
on them not being added to the device tree (due to the same value of
ID_FS_UUID).

Resolves: rhbz#615987
Related: rhbz#595866