Display the method selection details dialog on boot.iso (#620746).
This only applies to booting with boot.iso.  Without this patch, you'll get
the method selection screen where you can select the URL or NFS method, but
you'll never get the screen where you enter the details for finding the
package set.  This is because we already have stage2 data from mounting the
install.img off boot.iso.

With this patch, you'll get the method details dialog.  However, this has one
serious side effect.  The install.img used will be fetched from the remote
location, not the one on the boot.iso.  This means if you are doing a URL
install, you'll have to download the install.img anyway.

This seems to me to be the lowest risk change given the complexity of the
code and where we are in the schedule.