Check for ARPHRD_ETHER and ARPHRD_SLIP types in getDevices (#596826)
In the getDevices() function in isys, we check the type of the device
during discovery to see if the type is ARPHRD_ETHER.  If it isn't, we
skip that device when gathering network devices.  This filters out lo,
sit0, and other devices we don't want in the network device list.  But,
it does skip over one ethernet device type that is initialized as
ARPHRD_SLIP on s390.

(This is cherry-picked from rhel6-branch, but since the functionality is
already on master, I just made sure the test is using the ARPHRD_*
constants rather than 1 and 256.)