Do disklabel handling for whole disk formats unknown to anaconda (#619721)
Currently handleUdevDiskLabelFormat() skips disklabel handling if blkid
has found some sort of whole disk format, even when the type of whole disk
format is unknown to anaconda.

This seems wrong, and it causes bug 619721, which is about Firmware RAID
member disks being seen as a disk with an unknown whole disk format, rather
then a partitioned disk when anaconda is told to ignore Firmware RAID
metadata through the nodmraid cmdline option.

This patch also removes a similar check further down in
handleUdevDiskLabelFormat() which checks for the same thing, and thus is
redundant.