Remove the method selection block from the beginning of doLoaderMain.
This block was only there because we needed to figure out where the stage2
image was if you booted off a boot.iso or passed some magic parameter.  The
result of this block was a decision as to whether to display the askmethod
UI or not.

With the guarantee that however you boot, you have the complete installation
environment, all this fragile code is no longer necessary.  We simply display
the askmethod UI only if the user has passed "askmethod".  Otherwise, it is
assumed they want to use the default Fedora repos.  repo= kind of bypasses
this assumption but it doesn't affect UI display.