Create Size class for specifying device and fs sizes.
Created a new class called Size under the storage module.  It can
represent sizes where we have traditionally used megabytes.  The size is
stored as bytes in the object, but can convert to any other type unit
supported by the class.  The class inherits Decimal and can be used like
other Decimal objects (e.g., in mathematical expressions).

The class also has the humanReadable() method, which will output the
size in a human readable format to a specified number of decimal places,
or fewer if that makes sense in the given situation.

The initialization method also accepts a number of bytes or a string
specification for the size, such as "32 GB" or "640 kb".