Run anaconda in fullscreen mode (#663411).
This involved updating the main glade file, and I did so in glade-3
which makes the commit look much more involved than it needs to be.  I
also removed some container widgets that have basically been unused
since we got rid of the release notes viewer and the help panel.