Add logpicker tool into utils
The Logpicker is a simple utility for log reporting.
It obtains a dump from the Anaconda + some another information
and creates a compressed archive. Then it's able to send the archive to:
- Bugzilla (in case of Fedora)
- Red Hat Ticketing System (in case of RHEL)
- email
- FTP server
- another computer via SCP
- local filesystem

The dump from the Anaconda is obtained by the signal SIGUSR2.
Other information gathered by the Logpicker are:
 - Filesystem structure image (list of directories and files in ram disk).
 - Output from commands "dmsetup ls --tree" and "dmsetup info -c".
The Logpicker is easy to extend. So you can easily add your own log gathering
classes.

The main advantage of this utility is that all important logs can be easily
sended (reported) by only one command without break an instalation.

Example of use:
logpicker --bugzilla --login=bzlogin --idbug=999999