Check for and complain about package scriptlet errors (#531599).
We don't get a lot of information to go on from rpm/yum so just do the
best we can.  On interactive installs, complain that there was an error
and redirect the user to install.log where the error will have been
logged.  On kickstart installs, just log the complaint.