Attempt to bring the network up before saving a bug report (#635821).
Now that we're using report, we don't have quite the level of integration
necessary to only ask to bring up the network before saving to bugzilla
or scp.  Therefore, we just need to ask for the network before saving at all
and trust that report will fail correctly should the user not bring up an
interface but still try bugzilla.