Make sure we have the container before the member arrays. (#642765)
The multipath identification function randomizes the ordering of the
list of devices to scan, and commit be910ee4b0288d48c introduced a second
call to storage.reset. md-controlled biosraids will already be active at
the time of the second reset (after autopart_type) so the various md
devices will be subject to unpredictable scan order. This patch makes an
attempt to get the container into the tree on demand in the case where
the member arrays are scanned before the container.