use different approach to tweak gconf settings in the image (#642358).
Previously we used an .xml file. That bore two problems:
a) the xml contained paths to schmas making the tweaking code longer and
error-prone when the schema paths change.
b) the syntax we used to call gconftool-2 is not really documented in
gconftool's manpage.

The patch calls gconftool-2 directly, with specific paths, exactly in the
same way livecd spin kickstarts do.