Allow importing product.py in places where you won't have a .buildstamp.
If you're running anaconda through some external testing system, you're not
going to have a .buildstamp file.  However, the defaults dict for the
ConfigParser object doesn't allow for putting those defaults into a section
so all the later lookups will fail due to lack of a Main.  Moving to a big
block of .set calls fixes this up.