On upgrades, inform the user what action is taking place (#493249).
rpm doesn't tell the callback very much on upgrade cleanups.  We don't know
how many packages it's going to deal with.  We don't know the size of the
package being cleaned up.  We certainly don't know how long it's going to
take.  About the only thing we do know is what package is being processed.