Use libarchive helper functions in explodeRPM()
Simplify the archive extraction loop in explodeRPM() by using the
unpack.c helper functions.  Add a new parameter to explodeRPM, the
destination to unpack the archive to.  This was previously handled
by dlabelUnpackRPMDir() directly, but it made more sense to me to
move it in to the libarchive unpack helper.