Return mount's actual error codes instead of obfuscating them.
In addition to being unnecessary, this obfuscation does not account for
the fact that the codes can be combined using a bitwise or.

This is part of the reason we were unable to identify the actual problem
with the series of bugs involving the error "SystemError: (2, None)"
during F11, F12, and perhaps after that.

Related: rhbz#548592