Add pyanaconda/.libs to the PYTHONPATH for pylint.
This is to fix an error about not being able to import xutils when running
under pylint.  These fixes may seem a little arbitrary, but they're not.  If
the code says "import blah" where blah is a .so, we can add a directory to the
PYTHONPATH.  This is because blah.so will be in .libs/.

If the code says "from pyanaconda import blah", then we need to apply another
fix as the .libs/ directory is not involved.