Restore list-harddrives output to what users expect (#654436)
The list-harddrives command was changed to list partitions as well as
the whole disks.  Disks would only be shown if they had a valid
disklabel.  This is incorrect.  The command is meant to show what disks
are in the system and their size, usually so people can set up kickstart
partitioning commands or partition the disk on their own.

This patch changes the output from:

    /dev/sda 61057
    /dev/sdb 476940
    /dev/sdc 30524
    /dev/sda1 300
    /dev/sda2 60751
    /dev/sdb1 476937
    /dev/sdc1 30520

To:

    sda 61057.3359375
    sdb 476940.023438
    sdc 30524.0

Which is the way the command has worked in previous releases.