Create a MainframeDiskDevice class for common s390 attributes.
Create a MainframeDiskDevice class to hold common s390 attributes
across the disk types specific to that platform.  For now, it
does not do much except return self.desc for chreipl failures.
Oh, and build self.desc for DASDDevice and ZFCPDiskDevice which
is a slightly more human-readable string explaining where this
particular device lives on the system in z-speak.

Related: rhbz#632325