Fix looking up storage device IDs when writing out anaconda-ks.cfg (#591713).
Before, we were just writing out the UUID on the format objects.  However, the
object may not have a UUID when it's instantiated and we never go back and set
the attribute once we make filesystems.  So, we need to go and do a lookup at
the end when writeKS is done.  Using the major and minor device numbers is
unique enough for this purpose.