Adjust main window size based on install type (#667566)
Without the resize flag the live install would shrink the
main window to fit each screen's widget set. With resize=True
it will expand to fit the content, but not shrink. This also
allows maximize() to work and we don't need to set a specific
window size.

Here are the current screen sizing rules:

* Normal Install
  - Full screen
  - Desktop hint (no window decorators)
  - Not resizable
  - If the screen is < 600 the header is removed
* LiveCD
  - Maximized window
  - Resizable
* Forwarded X
  - Not maximized
  - Resizable