Support enabling repos listed but disabled in /etc/yum.repos.d (#663992).
Without this, there is no good way to enable those repos.  Adding a new repo
line with the same name and the baseurl leads to an error about duplicate
repos.  The only way is to add a new repo line with a made up name and the
baseurl.

With this, you can just do "repo --name=updates" and anaconda will enable the
repo found in /etc/yum.repos.d with whatever configuration it already has.