Provide a new mpath devicelib interface that does not reorder the devices.
This makes DeviceTree.populate()'s task simpler by providing the devices
in the same order udev reports them so we do not have to iterate through
the device list so many times (and force skipping some of the devices
during the first go).

I also hope the methods' input/output is clearer now.