Make sure /boot is mapped to a single LVM PV on s390x
When /boot lives on a logical volume on s390x, zipl requires that it map
to single physical volume.  This patch is a followup to
39244f8f49ee3864edf8bd54f41a6a37864ed56d to ensure that /boot is on a
logical volume that maps to a single physical volume.  I do that by
generating a list of parents, sorting, and taking the last one.  Not
exactly foolproof, but it should work for the autopart case which is the
main request.

Related: rhbz#618376

(based on 1a2d1305cb1d010d2074c763d7ab1524a116310d)