Make singlePV a more useful boolean, clean up _getSinglePV()
Followup again to the "/boot on LVM" commits for s390x.  The feature
request itself is odd, but it's close to a morbid curiosity now as to
whether or not it'll actually work.  The main limitation with /boot on
LVM on s390x is that the logical volume must be tied to a single
physical volume in the volume group.  Combined with the sorting change
commit, the patch does the following:

1) Makes singlePV a boolean attribute on the PartSpec and
   LVMLogicalVolumeDevice objects.

2) Key on the singlePV attribute rather than the mountpoint so the
   support for logical volumes on a single PV is a little more generic.

3) Modify _getSinglePV() to read the output of the vgs(8) command and
   find a PV with enough free space on it for the logical volume we are
   creating.  Once we find one, return it.  If we don't find one, raise
   an ugly exception and explain things to the user.

Now, the big gap here is what happens if there are no physical volumes
of at least 500MB, since we are limiting this singlePV stuff to /boot on
s390x?  The answer is the user can get the exception described in #3 and
do manual partitioning.  I personally have never seen a DASD under 2GB,
but I'm sure it's possible.  I don't really care.  I've fulfilled the
morbid curiosity of this request and know it works.

Related: rhbz#618376