Ensure ppc64 ramdisk.image file is <= 32MB (#675685)
Remove some kernel modules and firmware files so the ppc64
ramdisk.image.gz file is 32MB or less.  The core issue here is the TFTP
client code in on most (or all) ppc64 systems cannot transfer a file
larger than 32MB.

NOTE: This is not an ideal solution to a firmware bug.  ppc64 systems
from IBM are incapable of pulling over a file larger than 32MB via TFTP,
which prevents network booting on ppc64 when the install image grows
beyond that size.  These removals are a temporary workaround until IBM
can fix the firmware on affected systems.  The kernel modules and
firmware selected for removal is based on the assumption that most ppc64
users will not be using these devices during install time (or ever).  If
you do use devices that require these files during installation, we urge
you to seek another architecture or ask that you request a firmware fix
from IBM.