Make the dispatcher call the shots.
Changes the way the dispatcher and anaconda.intf cooperate in an attempt
to take control away from the interface and give it to the dispatcher
where it should be.

The dispatcher.moveStep() method is rewritten into a form that does not
include the complex condition in the while cycle. Also the method was
renamed to dispatch() to reflect that it (exclusively) containts the
step-shifting logic and that upon a single call to it the installation can
progress more than one step (or all of them).

For GUI interface implementation, this patch removes icw.setScreen and
isolates the code that creates next GUI screen into the display_step()
method. Similar changes happened for the text and command line interfaces.