Don't fail on missing %includes during loader kickstart processing (#676940).
Lots of people rely on the fact that a %pre script can generate a file that will
later be %included.  Proper kickstart processing involves first making a pass over
the file looking for any %pre scripts, ignoring all %include lines.  Then another
pass is made that handles all the %includes, which will pick up and that were
generated.

With the pykickstart integration, I made it so reboot and other loader-handled
kickstart commands can be in a %include file.  However, this also made a missing
%include a fatal error, which breaks the %pre-generated include trick.

This fix makes it so any %includes that do exist will be processed in loader, but
missing ones will not be a fatal error.  It will only be much later during
anaconda proper that a missing include will cause an error.