Do all dm handling inside addUdevDMDevice. (#672030)
The by-uuid lookup was returning a loop device since it has
the same contents as the dm device, which led to us not ever
adding live-osimg-min to the device tree for live installs.
Since UUIDs are not necessarily device-specific (as evidenced
by the case of fs UUID on loop device and same fs UUID on dm
device representing the contents of said loop device) we
can't rely on them for dm device lookups unless we
differentiate between device UUIDs and formatting UUIDs.