Fix problem with /boot on LVM on a zFCP multipath device
Followup to previous fixes for /boot on LVM for s390x.  When the
underlying partitions are on a multipath volume, the device found by
zipl is a dm-X device.  The code correctly picks this up so we can pass
it to chreipl, but we also try to run getDeviceByName() on the dm-X
name, which is unnecessary here.  When getDeviceByName() returns None,
return a 2-tuple of (dm-X, None) to the caller so we use dm-X as-is.

Related: rhbz#670856