Pass createUser and createGroup an arguments dict.
This makes the calls to these two functions a whole lot cleaner - instead
of passing an enormous list of arguments, I can just convert the pykickstart
objects into a dict and pass those in as the kwargs.

This sets up for passing more arguments in the future (for stuff like password
controls) without having to make the method calls even larger.