Fix test for resized LV to ensure we schedule the format resize action.
The currentSize attribute is only accurate for devices that have been
set up at least once. All existing devices in the devicetree are set up
during populate, but those in the temp vg we create for the lvm dialog
are not. So, to get the current size of the LV, we look to origlv,
which is the one in the devicetree.

Resolves: rhbz#681597