Handle boot loader upgrades on s390 (#682783)
Fix up the boot loader config reading and handling for upgrades on s390
so it works like x86 a little more.  We find an existing zipl
installation and ask the user what steps to take.  This results in the
boot loader configuration being updated correctly, reinstalled if
necessary, and the reIPL step succeeding.

Also fixed the missing 'rc' variable that was causing a traceback due to
the above code blocks missing.  So in instances where we cannot figure
out the boot loader configuration, do not traceback but just fall
through and complete the installation.  reIPL is extra stuff anyway.