Add transifex-client configuration file.
Translation work has changed again.  Rather than translators pushing
changes to our central repo, we are now responsible for pulling the
latest .po files from them.  We do this using the 'tx' command from the
transifex-client package.  The .tx/config file defines the
transifex-client settings for this project, which is (a) where
translations live, (b) the location of the pot file, and (c) the source
language.