Clean up display of free space in partitioning gui.
Show free regions within extended partitions as such in the graphic
as well as in the tree view.

Take advantage of the fact that free space "partitions" within an
extended have the PARTITION_LOGICAL flag set.

Lastly, use the grain size of the kernel-supplied alignment as the
minimum size of free regions that get displayed. Within the extended
we double this because the way logical partitions' metadata areas
are allocated leads to more conservative partition alignment.

Resolves: rhbz#626025