Determine existing md arrays' metadata version. (#731266)
This is needed if we are to validate boot arrays' metadata
version.

Also, don't validate metadata version for arrays we have not
yet created, as the metadata munging happens just before we
create the arrays.

(based on cherry pick of 9599b145793f3aa9e151f28c5cd2cda6f8dcd668)