Clarify display of free space in storage tree view.
Show free regions within an extended partition as children of the
extended. Also, use the grain size of the kernel-supplied alignment
as the minimum size of free regions we display. We double that for
free regions within an extended due to the great deal of variance
in sizes of logical partition metadata areas allocated by parted.

Resolves: rhbz#626025