Write OPTIONS for s390x net devs on non-network install methods (#649301)
Make sure the LAYER2 and PORTNO values collected by linuxrc.s390 are
translated to an appropriate OPTIONS line for the ifcfg-DEVICE file on
the target system.  This happens during network install methods because
loader converts those values in to an OPTIONS line, but if you use the
hard drive install method, the values are never translated to OPTIONS.
Do the translation in network.py to ensure we get those values on all
install methods.