Fix shlex traceback if pxeboot cmdline ends in a quote (#500198)
There's a pxelinux bug where it fails to add the beginning quote
when it tries to add BOOT_IMAGE= to the commandline and the label
has spaces (you get e.g.: 'linux BOOT_IMAGE=vmlinuz with label"')

Previously we tried to fix this up by adding a quote after
BOOT_IMAGE= whenever the cmdline ends in a quote. This causes us
to have unbalanced quotes if the cmdline is something like:
    linux BOOT_IMAGE=vmlinuz ethtool="speed 100 duplex"

With this change, we now split the commandline at BOOT_IMAGE= and
check to see if the right side has unbalanced quotes. If so, we add
a quote after BOOT_IMAGE=.