Set mainWindow size request to current res reported by xrandr (#694760)
Certain displays do not respond well to the maximize() calls on the mainWindow.
On these particular Xen guest installs, the user is unable to see the entire
anaconda window because the xrandr reported current resolution is 800x600, but
we're still trying to size it up to 1024x768 or higher.

When we scrape the current resolution from xrandr -q, and pass the width and
height to set_size_request(), the main anaconda window fills the whole display
the user is looking at.  I've tested it on various install methods at different
screen resolutions and it seems to work fine.  The only downside I see to this
is scraping the output of xrandr.  In the event that the output of xrandr -q is
empty or contains unexpected values, we don't do set_size_request().  The
default width and height on mainWindow is set to 800x600 anyway, so most
install cases will still work fine without a set_size_request() call.