Rework bootloader constraint checking routines.
Methods to validate stage1 and stage2 devices still return a
boolean value, but now also maintain a list of errors and warnings
found during validation which can be retrieved by the caller
afterwards. This allows us to simplify the Platform.checkBoot*
methods and remove some special-case code that checked for GPT
disklabels on non-EFI hardware.

New stage1 devices are no longer required to be valid prior to
setting since we need to do all validation on demand in order to
be able to present the user with a list of errors and warnings.

All checks that were previously done in Platform are now present
in BootLoader.