dispatch: clean up step skipping manipulations in kickstart.
There is no reason to postpone the processing of step skipping/enabling now.

Steps remember their states now and it is not possible to schedule a
skipped step or, conversely, make a step skipped after it has been
requested to run. Previously it was not possible to request a step because
the following events could as well mark it skipped again, making one
unsure whether or not a particular step will be run. Also previously
trying to 'unskip' a step after it has been marked as permanently skipped
failed silently which was also undesirable.

Because of those new dispatch features, there is a guarantee that once
kickstart command decides it has fulfilled the purpose of a particular
step and marks it as skipped, the step is never run.