dispatch: Fix rules for running the partitioning step.
This step is not scheduled by default, it is skipped for any text
installation and many kickstart commands (like autopart), it is requested
if the user selects "review partitioning layout" on the parttype screen.

There is a mechanism in autopartitioning that enables manual partitioning
screen (if available) when autopartitioning fails. Instead of checking the
availability this just tries to request the manual partitioning step and
catches DispatcherError for those cases where this is not possible.