In kickstart, specify multipaths by their wwids.
Aliases like 'mpatha' and 'mpathb' can cange between boots (or when new
devices are connected/disconnected from the system), just like with 'sda'
and 'sdb'. What is called for is being able to target a specific multipath
by its ID. This is achieved by allowing the user to specify a member
device of a multipath ('sda', or through /dev/disk/by-id/) and then
promoting it to the multipath device.

For example, commands like:

ignoredisk ... --only-use=disk/by-id/scsi-<wwid>
part ... --ondisk=disk/by-id/scsi-<wwid>

will make the mpath and the member drives with wwid <wwid> listed in
exclusiveDisks and create a partition on the same mpath.

Resolves: rhbz#677263
Related: rhbz#709653