Include all kernel crypto modules (#703782)
Brute force the module list and just include everything in the kernel
module crypto/ subdirectory.  I have no idea which ones are needed at
any given point in time and we periodically hit problems where the
kernel is expecting to find a module and we don't have it.

This list was compiled by unpacking kernel-2.6.18-262.el5.ppc64.rpm and
running:

    for f in *.ko ; do echo "$(basename $f .ko) ; done | sort | uniq

In the lib/modules/2.6.18-262.el5/kernel/crypto subdirectory.