Use xz compression on initrd.img, add back gfx firmware (#703914)
Use xz compression on the initrd.img file and bring back the explicitly
removed graphics hardware firmware.  These were eliminated on ppc64 to
get the size of the initrd.img down to 32MB or smaller due to
limitations of that platform.  With xz compression, we can hopefully
bring back the graphics firmware which are necessary for certain
systems.