Remove hasFreeDiskSpace and related code.
This function is needless complication. If you only check for free
regions greater or equal to 100MB people will be blocked from
creating new PReP or BIOS BOOT partitions even though they should
not be. OTOH if you only check for 1MB of free space that is of
nearly no informational value when creating a new LVM PV as
compared to having no information at all.