Delete the Mocked pyanaconda.product to fix product tests.
Since anaconda modules are such a tangle, running one of the earlier tests
causes pyanaconda.product to get imported early.  On that import, it gets
created as a Mock object, so the later import doesn't do anything useful and
all the product tests will therefore fail.  If we delete and then re-import,
everything will work.

This is only seen if you run everything in tests/, not if you just run the
product_test.py by itself.