Allow DriverDisc to contain vendor provided tools (#659790)
This extracts all files from /(s)bin, /usr/(s)bin, /lib(64) and
/usr/lib(64) into /tmp/DD hierarchy and adds those dirs to
PATH and LD_LIBRARY_PATH (in init).

The package has to Provide: installer-enhancement which version
matches the running anaconda's version to be considered for this
feature.

We implemented this feature to enable vendors to add their
tools for initializing special hardware. Our binaries and libraries
have higher priority, so in case of conflicts, anaconda wins.
This should prevent breakages to the installer environment.