Remove definite articles in the bootloader translation strings.
The following sentence is hard to translate:

_("The %s is huge.") % object , where object is taken from dictionary or an
array somewhere, e.g. [_("dog"), _("bird")]. In some languages
(Danish for instance) the article is sometimes part of the object word and
sometimes not, so depending on the article 'dog' could be either

'hunden' or 'en hund'.

In those cases the translater doesn't know how to translate the original
sentence without looking at the code. We shouldn't be composing sentences
in this way and avoid using definite/indefinite articles where we do. (The
only perfect solution though is to have all possible sentences listed and
not construct them).

A similar problem arises e.g. with word order that changes among languages
and capital letters.

Resolves: rhbz#713278