Reuse existing /boot/efi and ignore vfat on non-boot devices (#723976)
Set the size of new /boot/efi partition requests to the minimum size
instead of a size that is too small.
Use the drivelist to set the device to use for pre-existing /boot/efi
partitions.
Only one /boot/efi is needed on the boot device, set it when
skipping unneeded ones.

Resolves: rhbz#723976
Resolves: rhbz#729062