Copy /etc/multipath/wwids to the sysimage.
/etc/multipath/wwids is a list of activated multipaths generated by the
multipath tools. If it is missing in the sysimage it will be missing in
the dracut initramfs and that, in rare cases, can cause race between mpath
and lvm during boot.

This is a merge froh rhel6-branch but includes a unit tests of the new
iutil.copy_to_sysimage().

Resolves: rhbz#701371