Do not invalidate _cachedDevices (#618535)
If _cachedDevices is invalidated, singlepath devices get re-added to the
UI for each new advanced storage device you add.  So, with this change,
_cachedDevices will end up storing the first path of a multipath device
even after it's determined to be a multipath device.  Seems to work fine
as the _cachedDevices list is only for internal use and leaving it in
the cache doesn't hurt anything since a device by that name won't show
up later.