Fix the partitioning section of interactive kickstart installs (#641861).
Note that behavior around going back from the examining UI to the part
type UI can be a little weird.  If you have a kickstart file with partition
commands in it and you go back, those commands will be forgotten.  This is
because we have no way of reloading the kickstart data short of reexecuting
the entire file (which is too messy).

On the other hand, it's kind of what you'd expect - if you go back from any
step, your choices are blanked out so you try again from scratch.