Require BIOS boot partition for GPT bootdisk on BIOS systems.
Changes the warning to an error and also fixes two other bugs.

First, it fixes an unbound recursion issue introduced by
commit e8ae9031f203b15. The check for stage2 device being the
stage1 device doesn't make sense in _gpt_disk_has_bios_boot.
If the stage1 device is not a disk that method will not
require a bios boot partition anyway.

Second, the bios boot check was being run regardless of
whether the platform requires it. Checking the weight is the
only method I know of to determine this.

It also simplifies the error message so fewer people get
confused by all the acronyms.