iutil: make getArch() return ppc64 on ppc64 (#736721)
In the past, yum considered ppc64's basearch to be 'ppc', but in F16
ppc64 basearch is 'ppc64'. This makes images.findFirstIsoImage() fail,
since lorax writes the basearch into .discinfo.

This changes the behavior of getArch to return ppc64 on ppc64, which
will allow ppc64-based installs to continue.

Note that this means that a ppc64 system will reject an image that has
'ppc' as its basearch. People intending to build hybrid 32/64 bit media
will have to find some way to work around that.