Make sure there are no tempvg paths even if formatting. (#737916)
If the user tried to reuse a rootfs they got the device attr set
to point to the tempvg but because of us forcing them to they are
reformatting it. In this scenario we have to force the device
attr back to point to the real device node.
(cherry picked from commit 0cceef0efad5927ff13ae47abcc3e5ecd35b8c6c)