Do not reactivate network device needlessly on s390 (#739846)
On s390, the device is activated when NM is started due to ifcfg file
generated by linuxrc.390. There is no need to reactivate it with
the same parameters (it would even lead to fail).

This condition (2nd part) from kickstartNetworkUp was causing
(or perhaps to put it better - not preventing) the ractivation in
case of already having been connected:

    if ((is_nm_connected() == TRUE) &&
        (loaderData->netDev != NULL) && (loaderData->netDev_set == 1))
        return 0;

To satisfy it, just set the values properly in readNetInfo (the patch).