Show cleardisks gui always to allow selecting a boot disk. (#744088)
This is not ideal but for F16 it accomplishes the goal of forcing
boot device selection while only using existing screens in existing
order. For custom, all disks are initially on the right and the
add/remove buttons are disabled.
(cherry picked from commit 684db1909b382dc225084f8e6e53b970382eadaa)