Block button handlers after a button is pressed (#738217)
There's a race condition if the buttons are pressed almost
simultaneously, causing more dialogs to appear at the same time.
Although all the dialogs are modal, gtk doesn't make the buttons
inactive fast enough. The dialogs are not made to work this way
resulting in tracebacks. Instead of fixing the dialogs we need
to make sure this won't happen, because having more dialogs
open at the same time makes no sense.

This is the least invasive fix for this race condition that
I could come up with.