Change what the third column of lang-table means.
It now means "is text mode supported for this language?".  It used to store
which font we'd use to display the language, where "none" meant that text
mode didn't work in that language.  However, we don't ever set the font
based upon that column and we shouldn't really ever be setting the font
ourselves anyway.