Handle %posttrans RPMCALLBACK_INST_OPEN_FILE properly (#506361)
The first attempt to fix the bug has broken http:// and ftp://
installations. Also (independent of that) I've learned that the additional
call of RPMCALLBACK_INST_OPEN_FILE is due to %posttrans scriplet
(so no bug in yum).

Anaconda failed to handle this additional call in specific case of
case of nfs method and http additional repository.

The fail was caused by self.method.unlinkFilename(fn) in
RPMCALLBACK_INST_CLOSE_FILE being dependent on method (i.e. do nothing
for nfs, remove rpm file from cache for http). In the problematic
case the package of additional http repository was not removed from cache
because the install method is nfs. As a consequence, in the second
call of RPMCALLBACK_INST_OPEN_FILE repo.getPackage(po) failed to
download the package.

Related: rhbz#506361