loader/serial.c: correctly switch controlling terminal (#732302)
So after loader picks which terminal it likes best, it opens it, uses
dup2() to replace its fd (0,1,2) with the new terminal, and then does:
  setsid(); ioctl(0, TIOCSCTTY, NULL);
to switch its controlling terminal to the new terminal (on fd 0).

In commit b8886ea3, the code was refactored a bit and the ioctl() got
moved *above* the dup2(), which meant that fd 0 was still the *old*
terminal, so the ioctl() would have no useful effect.

This commit makes ioctl() operate on the correct fd, and changes NULL
to '0' (since ioctl takes an int arg, not a pointer).