Define exactly which arguments Spokes, Hubs, andand UserInterfaces may count on.
For now, I'm just letting them get ksdata, devicetree, and instclass.  More
may need to be defined later (packaging backend and networking, perhaps) but
the idea is to keep this list small and to have them be explicit in what they
need.  Also, the anaconda object will never be a parameter.  Don't even ask.